Created and Fixed the CRUD service for group feeds page. Tried to write test for controller but Couldn't figure out how to mock principal and authority token.

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/service/PostService.java
@@ -6,7 +6,7 @@ import nz.ac.canterbury.seng302.portfolio.model.entity.PostModelRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.lang.reflect.Array;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -15,25 +15,61 @@ public class PostService {
     @Autowired
     private PostModelRepository postRepository;
 
+    public List<PostModel> getAllPosts () {
+        return (ArrayList) postRepository.findAll();
+    }
     public List<PostModel> getAllPostsForAGroup(int groupId) {
         return postRepository.findPostModelByGroupId(groupId);
     }
 
-    public PostModel createPost(BasePostContract newPost, int userId) {
-        return postRepository.save(new PostModel(newPost.groupId(), userId, newPost.postContent()));
+    public boolean createPost(BasePostContract newPost, int userId) {
+        try {
+            if (newPost.postContent().length() == 0) {
+                return false;
+            }
+            PostModel postModel = new PostModel(newPost.groupId(), userId, newPost.postContent());
+            postRepository.save(postModel);
+            return true;
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return false;
     }
 
-    public void deletePost(int postId) {
-        postRepository.deleteById(postId);
+    public boolean deletePost(int postId) {
+
+        try {
+            var postFound = postRepository.findById(postId);
+            if (!postFound.isEmpty()) {
+                postRepository.deleteById(postId);
+                return true;
+            } else {
+                return false;
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            return false;
+        }
     }
 
-    public void updatePost(BasePostContract updatedPost, int postId) {
-        var post = postRepository.findById(postId).orElseThrow();
-        post.setPostContent(updatedPost.postContent());
-        postRepository.save(post);
+    public boolean updatePost(BasePostContract updatedPost, int postId) {
+        try {
+            var post = postRepository.findById(postId).orElseThrow();
+            post.setPostContent(updatedPost.postContent());
+            postRepository.save(post);
+            return true;
+        } catch (Exception e) {
+            e.printStackTrace();
+            return false;
+        }
     }
 
     public PostModel getPostById(int postId) {
-        return postRepository.findById(postId).orElseThrow();
+        try {
+            return postRepository.findById(postId).orElseThrow();
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 }
